move package json transformer to its own class and use from node modules linker as well
